Clean up product views and unify cards
