+++ NexStock1.0/View/InventoryCardView.swift
@@ -8,7 +8,7 @@
 import SwiftUI
 
 struct InventoryCardView: View {
-    let product: DetailedProductModel
+    let product: ProductModel
     var onTap: (() -> Void)? = nil
     @EnvironmentObject var theme: ThemeManager
     @EnvironmentObject var localization: LocalizationManager
